requirements.txt
flask
Jinja2
flask_pymongo
nsetools